change pod structure apps.image into images.app
