--- assets/app/protected/apps/template.hbs
@@ -1,3 +1,0 @@
-<div class="configuration section">
-  {{outlet}} 
-</div>
