+++ assets/app/protected/dashboard/template.hbs
@@ -15,7 +15,7 @@
       {{/link-to}}
     {{/card-component}}
     {{#card-component title="Published applications" content=apps.length clickOnContent="goToApps" wait=loadState.application}}
-      {{#link-to 'protected.apps'}}
+      {{#link-to 'protected.images'}}
         <span class='color-primary'>View apps</span>
       {{/link-to}}
     {{/card-component}}
