+++ assets/app/protected/files/notactivated/template.hbs
@@ -7,7 +7,7 @@
       This feature requires a first successful connection to an application to be available.<br />
       Click "Go to applications" to be redirected to the page were you will be able launch your apps
     </div>
-    {{#link-to "protected.apps"}}
+    {{#link-to "protected.images"}}
       <button class='btn btn-primary btn-block'>
         Go to applications
       </button>
