+++ assets/app/protected/histories/index/table/history-list/app-name/template.hbs
@@ -1,3 +1,3 @@
-{{#link-to 'protected.apps.app' record.connectionId class="truncate"}}
+{{#link-to 'protected.images.app' record.connectionId class="truncate"}}
   {{record.application}}
 {{/link-to}}
