+++ assets/app/protected/images/image/template.hbs
@@ -1,6 +1,10 @@
-<div class='content-wrapper'>
+<div class='section content-wrapper'>
   <h1 class='m-b-2'>
-    {{#link-to 'protected.apps'}}Applications{{/link-to}}
+    {{#link-to 'protected.images'}}Applications{{/link-to}}
+    <div class='in-bl'>
+      {{icon-component class="va-sub" materialIcon="navigate_next" size=30}}
+    </div>
+    Image
     <div class='in-bl'>
       {{icon-component class="va-sub" materialIcon="navigate_next" size=30}}
     </div>
