+++ assets/app/protected/images/index/controller.js
