+++ assets/app/protected/images/index/route.js
