+++ assets/app/protected/machines/index/table/machine-list/image/template.hbs
@@ -1,4 +1,4 @@
-{{#link-to 'protected.apps.image' record.image.id}}
+{{#link-to 'protected.images.image' record.image.id}}
   {{#text-component class='in-bl' content=record.image.id position="right"}}
     {{record.image.name}}
   {{/text-component}}
