+++ assets/app/protected/template.hbs
@@ -17,7 +17,7 @@
             </li>
           {{/if}}
           <li>
-            {{#link-to 'protected.apps'}}
+            {{#link-to 'protected.images'}}
               <span class="icon">
                 <i class="material-icons">apps</i>
               </span>
