+++ assets/app/protected/users/groups/group/images/template.hbs
@@ -13,7 +13,7 @@
           <tr>
             <td scope="row">
               <span class="link in-bl">
-                {{#link-to 'protected.apps.image' image}}
+                {{#link-to 'protected.images.image' image}}
                   {{image.name}}
                 {{/link-to}}
               </span>
